move domSource to Eval, as it's an Eval source
