+++ src/Eval.cpp
@@ -26,6 +26,7 @@
 
 #include <cmake.h>
 #include <Eval.h>
+#include <DOM.h>
 #include <map>
 #include <time.h>
 #include <Context.h>
@@ -34,7 +35,7 @@
 #include <shared.h>
 #include <format.h>
 
-extern Task& contextTask;
+extern Task* contextTask;
 
 ////////////////////////////////////////////////////////////////////////////////
 // Supported operators, borrowed from C++, particularly the precedence.
@@ -99,6 +100,19 @@ static bool namedConstants (const std::string& name, Variant& value)
     return false;
 
   return true;
+}
+
+////////////////////////////////////////////////////////////////////////////////
+// Support for evaluating DOM references (add with `e.AddSource(domSource)`)
+bool domSource (const std::string& identifier, Variant& value)
+{
+  if (getDOM (identifier, contextTask, value))
+  {
+    value.source (identifier);
+    return true;
+  }
+
+  return false;
 }
 
 ////////////////////////////////////////////////////////////////////////////////
@@ -299,10 +313,10 @@ void Eval::evaluatePostfixStack (
       else if (token.first == "^")        result = left ^ right;
       else if (token.first == "%")        result = left % right;
       else if (token.first == "xor")      result = left.operator_xor (right);
-      else if (token.first == "~")        result = left.operator_match (right, contextTask);
-      else if (token.first == "!~")       result = left.operator_nomatch (right, contextTask);
-      else if (token.first == "_hastag_") result = left.operator_hastag (right, contextTask);
-      else if (token.first == "_notag_")  result = left.operator_notag (right, contextTask);
+      else if (token.first == "~")        result = left.operator_match (right, *contextTask);
+      else if (token.first == "!~")       result = left.operator_nomatch (right, *contextTask);
+      else if (token.first == "_hastag_") result = left.operator_hastag (right, *contextTask);
+      else if (token.first == "_notag_")  result = left.operator_notag (right, *contextTask);
       else
         throw format ("Unsupported operator '{1}'.", token.first);
