+++ src/Filter.h
@@ -32,8 +32,6 @@
 #include <Task.h>
 #include <Variant.h>
 
-bool domSource (const std::string&, Variant&);
-
 class Filter
 {
 public:
